moved exceptions into separate folder
