+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/application/CheckInServiceImpl.java
@@ -5,6 +5,10 @@ import at.fhv.hotelsoftware.application.dto.RoomDTO;
 import at.fhv.hotelsoftware.domain.api.BookingRepository;
 import at.fhv.hotelsoftware.domain.api.RoomRepository;
 import at.fhv.hotelsoftware.domain.model.*;
+import at.fhv.hotelsoftware.domain.model.exceptions.BookingNotFoundException;
+import at.fhv.hotelsoftware.domain.model.exceptions.NotEnoughRoomsException;
+import at.fhv.hotelsoftware.domain.model.exceptions.RoomAlreadyOccupiedException;
+import at.fhv.hotelsoftware.domain.model.exceptions.RoomNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
