+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/application/ViewRoomServiceImpl.java
@@ -1,19 +1,16 @@
 package at.fhv.hotelsoftware.application;
 
 import at.fhv.hotelsoftware.application.api.ViewRoomService;
-import at.fhv.hotelsoftware.application.dto.BookingDTO;
 import at.fhv.hotelsoftware.application.dto.RoomDTO;
 import at.fhv.hotelsoftware.domain.api.RoomRepository;
-import at.fhv.hotelsoftware.domain.model.Booking;
 import at.fhv.hotelsoftware.domain.model.BookingId;
 import at.fhv.hotelsoftware.domain.model.Room;
-import at.fhv.hotelsoftware.domain.model.*;
+import at.fhv.hotelsoftware.domain.model.exceptions.RoomNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
 @Component
 public class ViewRoomServiceImpl implements ViewRoomService {
