+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/application/api/CheckInService.java
@@ -2,6 +2,10 @@ package at.fhv.hotelsoftware.application.api;
 
 import at.fhv.hotelsoftware.application.dto.RoomDTO;
 import at.fhv.hotelsoftware.domain.model.*;
+import at.fhv.hotelsoftware.domain.model.exceptions.BookingNotFoundException;
+import at.fhv.hotelsoftware.domain.model.exceptions.NotEnoughRoomsException;
+import at.fhv.hotelsoftware.domain.model.exceptions.RoomAlreadyOccupiedException;
+import at.fhv.hotelsoftware.domain.model.exceptions.RoomNotFoundException;
 
 import java.util.List;
