+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/application/api/CreateInvoiceService.java
@@ -1,9 +1,9 @@
 package at.fhv.hotelsoftware.application.api;
 
 import at.fhv.hotelsoftware.domain.model.BookingId;
-import at.fhv.hotelsoftware.domain.model.BookingNotFoundException;
-import at.fhv.hotelsoftware.domain.model.CustomerNotFoundException;
-import at.fhv.hotelsoftware.domain.model.InvoiceAlreadyCreatedException;
+import at.fhv.hotelsoftware.domain.model.exceptions.BookingNotFoundException;
+import at.fhv.hotelsoftware.domain.model.exceptions.CustomerNotFoundException;
+import at.fhv.hotelsoftware.domain.model.exceptions.InvoiceAlreadyCreatedException;
 
 public interface CreateInvoiceService {
     void createInvoice(BookingId bookingId) throws BookingNotFoundException, CustomerNotFoundException, InvoiceAlreadyCreatedException;
