+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/application/api/ViewBookingService.java
@@ -1,10 +1,7 @@
 package at.fhv.hotelsoftware.application.api;
 
 import at.fhv.hotelsoftware.application.dto.BookingDTO;
-import at.fhv.hotelsoftware.application.dto.RoomDTO;
-import at.fhv.hotelsoftware.domain.model.BookingId;
-import at.fhv.hotelsoftware.domain.model.BookingNotFoundException;
-import at.fhv.hotelsoftware.domain.model.Room;
+import at.fhv.hotelsoftware.domain.model.exceptions.BookingNotFoundException;
 
 import java.util.List;
