+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/application/api/ViewCustomerService.java
@@ -1,7 +1,7 @@
 package at.fhv.hotelsoftware.application.api;
 
 import at.fhv.hotelsoftware.application.dto.CustomerDTO;
-import at.fhv.hotelsoftware.domain.model.CustomerNotFoundException;
+import at.fhv.hotelsoftware.domain.model.exceptions.CustomerNotFoundException;
 import at.fhv.hotelsoftware.domain.model.CustomerId;
 
 public interface ViewCustomerService {
