+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/application/api/ViewInvoiceService.java
@@ -2,7 +2,7 @@ package at.fhv.hotelsoftware.application.api;
 
 import at.fhv.hotelsoftware.application.dto.InvoiceDTO;
 import at.fhv.hotelsoftware.domain.model.BookingId;
-import at.fhv.hotelsoftware.domain.model.BookingNotFoundException;
+import at.fhv.hotelsoftware.domain.model.exceptions.BookingNotFoundException;
 
 import java.util.List;
