+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/application/api/ViewRoomService.java
@@ -1,10 +1,8 @@
 package at.fhv.hotelsoftware.application.api;
 
-import at.fhv.hotelsoftware.application.dto.BookingDTO;
 import at.fhv.hotelsoftware.application.dto.RoomDTO;
-import at.fhv.hotelsoftware.domain.model.BookingId;
 import at.fhv.hotelsoftware.domain.model.Room;
-import at.fhv.hotelsoftware.domain.model.RoomNotFoundException;
+import at.fhv.hotelsoftware.domain.model.exceptions.RoomNotFoundException;
 
 import java.util.List;
