+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/domain/model/exceptions/BookingNotFoundException.java
@@ -1,4 +1,4 @@
-package at.fhv.hotelsoftware.domain.model;
+package at.fhv.hotelsoftware.domain.model.exceptions;
 
 public class BookingNotFoundException extends Exception {
     public BookingNotFoundException(String message){
