+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/domain/model/exceptions/CustomerNotFoundException.java
@@ -1,4 +1,4 @@
-package at.fhv.hotelsoftware.domain.model;
+package at.fhv.hotelsoftware.domain.model.exceptions;
 
 public class CustomerNotFoundException extends Exception {
     public CustomerNotFoundException(String message) {
