+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/domain/model/exceptions/InvoiceAlreadyCreatedException.java
@@ -1,4 +1,4 @@
-package at.fhv.hotelsoftware.domain.model;
+package at.fhv.hotelsoftware.domain.model.exceptions;
 
 public class InvoiceAlreadyCreatedException extends Exception{
     public InvoiceAlreadyCreatedException(String message) {
