+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/domain/model/exceptions/NotEnoughRoomsException.java
@@ -1,4 +1,4 @@
-package at.fhv.hotelsoftware.domain.model;
+package at.fhv.hotelsoftware.domain.model.exceptions;
 
 public class NotEnoughRoomsException extends Exception {
     public NotEnoughRoomsException(String message) {super(message);}
