+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/domain/model/exceptions/RoomAlreadyOccupiedException.java
@@ -1,4 +1,4 @@
-package at.fhv.hotelsoftware.domain.model;
+package at.fhv.hotelsoftware.domain.model.exceptions;
 
 public class RoomAlreadyOccupiedException extends Exception{
     public RoomAlreadyOccupiedException(String message){
