+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/domain/model/exceptions/RoomNotFoundException.java
@@ -1,4 +1,4 @@
-package at.fhv.hotelsoftware.domain.model;
+package at.fhv.hotelsoftware.domain.model.exceptions;
 
 public class RoomNotFoundException extends Exception{
     public RoomNotFoundException(String message){
