+++ hotelsoftware/src/main/java/at/fhv/hotelsoftware/view/BookingController.java
@@ -6,7 +6,7 @@ import at.fhv.hotelsoftware.application.dto.CustomerDTO;
 import at.fhv.hotelsoftware.application.dto.RoomDTO;
 import at.fhv.hotelsoftware.domain.api.BookingRepository;
 import at.fhv.hotelsoftware.domain.api.CustomerRepository;
-import at.fhv.hotelsoftware.domain.model.CustomerNotFoundException;
+import at.fhv.hotelsoftware.domain.model.exceptions.*;
 import at.fhv.hotelsoftware.domain.model.*;
 import at.fhv.hotelsoftware.view.form.FreeRoomListWrapper;
 import at.fhv.hotelsoftware.view.form.BookingForm;
